First Commit: Setup
This is everything set up and ready for work.
Bootstrap installed using nvm (so when downloading this repository you need to write "nvm install" in the terminal.
A sass folder and a scss file in it created ready for customizing Bootstrap using SASS.
Gitignore "Node" chosen when creating the repository.
Everything needed for bootstrap (bootstrap icons, bootstrap javascript) included in html index file.